Include CORS package and allow CORS requests
CORS (Cross Origin Resource Sharing) is required since this project hosts
the web app and API on separate domains. If the web app makes an HTTP
request to the API without this enabled, the browser will return
ERR_CONNECTION_REFUSED even though the request is valid. In this case we
simply allow all cross origin requests since the project will be hosted
locally, but in a real scenario where the app is deployed, we would
restrict the requests to our specific domain for the web app.